docs: document phase 11 release test execution features
- Add backend/docs/releases.md covering test sessions, accept-as-
  known-issue, Jira bug filing, per-test evidence, assignees, and
  the readiness rules.
- Expand backend/docs/manual-tests.md with groups, requirements
  traceability, and flakiness signals; list the new fields returned
  by the list/detail endpoints.
- Catalog migrations 024–029 in backend/docs/migrations.md.
- Add a Phase 11 entry to docs/roadmap.md marking the new release-
  grade test execution features as shipped.
- Extend the phase 9/10 smoke test with coverage for groups (bulk-
  link is idempotent), sessions (auto-complete, accept/revoke,
  failures-only rerun), and requirement link/unlink; tighten the
  sign-off test to use an explicit checklist so auto-rule items
  don't interfere. Swap the flakiness SQL cast from float to
  numeric to keep pass_rate exact.

+++ backend/src/routes/manual-tests.ts
@@ -55,9 +55,9 @@ router.get("/", async (req, res) => {
        ),
        flakiness AS (
          SELECT manual_test_id,
-                COUNT(*)                                          AS total_runs,
-                COUNT(*) FILTER (WHERE status = 'passed')::float  AS passes,
-                COUNT(*) FILTER (WHERE status = 'failed')         AS failures
+                COUNT(*)                                           AS total_runs,
+                COUNT(*) FILTER (WHERE status = 'passed')::numeric AS passes,
+                COUNT(*) FILTER (WHERE status = 'failed')          AS failures
            FROM release_test_session_results
           WHERE status IN ('passed', 'failed')
           GROUP BY manual_test_id
